refactor: rewrite temporary components
