+++ src/components/FlowManager.tsx
@@ -1,12 +1,12 @@
 import { StyleSheet, Text, View } from "react-native"
 import type { FlowManagerProps } from "../types/props"
-import { TemporaryTransitionComponents } from "../_core/components/TemporaryTransitionComponents"
+import { DisplayerTemporaryComponents } from "../_core/components/DisplayerTemporaryComponents"
 
 export const FlowManager = ({ children }: FlowManagerProps) => {
 	return (
 		<View style={[StyleSheet.absoluteFill]}>
 			{children}
-			<TemporaryTransitionComponents />
+			<DisplayerTemporaryComponents />
 		</View>
 	)
 }
